# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "restful-api-laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "get all users",
                "operationId": "a7821e1181a599ec4850700f74185e39",
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "create the user",
                "operationId": "b9ed5ffee044541198173eaf2aae77db",
                "parameters": [
                    {
                        "name": "user",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "example": {
                                "name": "john",
                                "age": 30
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "get the user",
                "operationId": "44b0d8a98660ee385eb420fac9b4720f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "delete the user",
                "operationId": "6f0b54081cc2c612a22b8d1abcdd1be5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "update the user",
                "operationId": "71cbd20ff9a6369b1ca32b89a8b6a89a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "example": {
                                "name": "john",
                                "age": 30
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        }
    }
}